# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive WhatsApp Business API built with Go, following Clean Architecture principles.\nProvides endpoints for session management, messaging, contacts, groups, media handling, and integrations.",
        "title": "zpwoot WhatsApp API",
        "contact": {
            "name": "zpwoot Support",
            "url": "https://github.com/your-org/zpwoot",
            "email": "support@zpwoot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sessions/{sessionId}/chatwoot/find": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Chatwoot configuration for a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatwoot"
                ],
                "summary": "Find Chatwoot Configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration found",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatwootResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Configuration not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/chatwoot/set": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create or update Chatwoot configuration for a session (upsert operation). To disable/remove integration, set enabled: false",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatwoot"
                ],
                "summary": "Set Chatwoot Configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chatwoot configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateChatwootRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatwootResponse"
                        }
                    },
                    "201": {
                        "description": "Configuration created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatwootResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/chatwoot/webhook": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Receive webhook events from Chatwoot and forward to WhatsApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatwoot"
                ],
                "summary": "Chatwoot Webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Webhook payload from Chatwoot",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatwootWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatwootWebhookResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "name is required"
                }
            }
        },
...
        "dto.ChatwootAccount": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootAttachment": {
            "type": "object",
            "properties": {
                "data_url": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "thumb_url": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootContact": {
            "type": "object",
            "properties": {
                "contact_inboxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatwootContactInbox"
                    }
                },
                "custom_attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootContactInbox": {
            "type": "object",
            "properties": {
                "inbox_id": {
                    "type": "integer"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootConversation": {
            "type": "object",
            "properties": {
                "agent_last_seen_at": {
                    "type": "integer"
                },
                "contact": {
                    "$ref": "#/definitions/dto.ChatwootContact"
                },
                "contact_last_seen_at": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "custom_attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "integer"
                },
                "inbox": {
                    "$ref": "#/definitions/dto.ChatwootInbox"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatwootMessage"
                    }
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "dto.ChatwootInbox": {
            "type": "object",
            "properties": {
                "channel_type": {
                    "type": "string"
                },
                "enable_auto_assignment": {
                    "type": "boolean"
                },
                "greeting_enabled": {
                    "type": "boolean"
                },
                "greeting_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "working_hours_enabled": {
                    "type": "boolean"
                }
            }
        },
        "dto.ChatwootMessage": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatwootAttachment"
                    }
                },
                "contact": {
                    "$ref": "#/definitions/dto.ChatwootContact"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "conversation": {
                    "$ref": "#/definitions/dto.ChatwootConversation"
                },
                "created_at": {
                    "type": "integer"
                },
                "echo": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "inbox": {
                    "$ref": "#/definitions/dto.ChatwootInbox"
                },
                "message_type": {
                    "type": "integer"
                },
                "private": {
                    "type": "boolean"
                },
                "sender": {
                    "$ref": "#/definitions/dto.ChatwootUser"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "autoCreate": {
                    "type": "boolean"
                },
                "convPending": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "ignoreJids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "importContacts": {
                    "type": "boolean"
                },
                "importDays": {
                    "type": "integer"
                },
                "importMessages": {
                    "type": "boolean"
                },
                "inboxId": {
                    "type": "string"
                },
                "inboxName": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "mergeBrazil": {
                    "type": "boolean"
                },
                "number": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "reopenConv": {
                    "type": "boolean"
                },
                "sessionId": {
                    "type": "string"
                },
                "signDelimiter": {
                    "type": "string"
                },
                "signMsg": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "webhookUrl": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootUser": {
            "type": "object",
            "properties": {
                "availability_status": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.ChatwootWebhookRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/dto.ChatwootAccount"
                },
                "changed_by": {
                    "$ref": "#/definitions/dto.ChatwootUser"
                },
                "contact": {
                    "$ref": "#/definitions/dto.ChatwootContact"
                },
                "conversation": {
                    "$ref": "#/definitions/dto.ChatwootConversation"
                },
                "event": {
                    "type": "string"
                },
                "inbox": {
                    "$ref": "#/definitions/dto.ChatwootInbox"
                },
                "message": {
                    "$ref": "#/definitions/dto.ChatwootMessage"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "dto.ChatwootWebhookResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CreateChatwootRequest": {
            "type": "object",
            "required": [
                "accountId",
                "token",
                "url"
            ],
            "properties": {
                "accountId": {
                    "type": "string",
                    "minLength": 1
                },
                "autoCreate": {
                    "type": "boolean"
                },
                "convPending": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "ignoreJids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "importContacts": {
                    "type": "boolean"
                },
                "importDays": {
                    "type": "integer",
                    "maximum": 365,
                    "minimum": 1
                },
                "importMessages": {
                    "type": "boolean"
                },
                "inboxId": {
                    "type": "string"
                },
                "inboxName": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "mergeBrazil": {
                    "type": "boolean"
                },
                "number": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "reopenConv": {
                    "type": "boolean"
                },
                "signDelimiter": {
                    "type": "string"
                },
                "signMsg": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string",
                    "minLength": 1
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication. Use the value from .env file (API_KEY variable). Send as 'Authorization: your-api-key' (without Bearer prefix)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}